requirements.txt
gTTS==2.2.3
moviepy==1.0.3
python-telegram-bot==13.15
imageio[ffmpeg]
requests
numpy
Pillow
python-dotenv
